pyproject.toml
[project]
name = "book-data-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "altair<5",
    "duckdb>=1.4.4",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.133.0",
    "numpy>=2.4.2",
    "openai>=2.23.0",
    "pandas>=3.0.1",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.19.0",
    "uvicorn>=0.41.0",
]
